compiler: Move Result methods to codegen
The compiler shouldn't know anything about individual languages. The
compiler package no longer imports the codegen package. The
*compiler.Result struct isn't a perfect interface, but it's much better
than the Generatable interface.
