+++ internal/opts/parser.go
@@ -1,0 +1,5 @@
+package opts
+
+type Parser struct {
+	UsePositionalParameters bool
+}
